% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTBonusFromQual.R
\name{MTBonusFromQual}
\alias{MTBonusFromQual}
\title{A convenience function to apply bonuses based on qualification scores.}
\usage{
MTBonusFromQual(HITTypeId = NULL, workerIds = NULL, assignmentIds = NULL,
  results = NULL, quals = NULL, bonusThresholds = NULL,
  bonusAmount = NULL, reason = "Thank you for completing these HITs!",
  sandbox = TRUE, verbose = FALSE, confirm = TRUE)
}
\arguments{
\item{HITTypeId}{A single HITTypeId. \code{"HITTypeId", "workerIds", and "assignmentIds"}
XOR \code{"results"} must be specified.}

\item{workerIds}{A single WorkerId or vector of WorkerIds. \code{"HITTypeId", "workerIds", and "assignmentIds"}
XOR \code{"results"} must be specified.}

\item{assignmentIds}{A single AssignmentId or vector of AssignmentIds. If specified, must be the
same length as \code{"workerIds"}. \code{"HITTypeId", "workerIds", and "assignmentIds"}
XOR \code{"results"} must be specified.}

\item{results}{A data.frame returned from assignments returned by MTurk. \code{"HITTypeId", "workerIds", and "assignmentIds"}
XOR \code{"results"} must be specified.}

\item{quals}{A single QualificationId or vector of QualificationIds.}

\item{bonusThresholds}{A single bonus threshold or vector of bonus thresholds. Must be the same
length as \code{quals}. The threshold means the worker must have a value greater than or equal
to the threshold amount. Must be \code{numeric}.}

\item{bonusAmount}{The amount of money, in dollars, to be given to each worker.}

\item{reason}{The reason to send the workers for receiving the bonus.
Default is \code{"Thank you for completing these HITs!"}.}

\item{sandbox}{Default is \code{TRUE}. Defines whether to work on live or sandbox Mechanical Turk site.}

\item{verbose}{Defines whether additional output is printed to the console.}

\item{confirm}{If \code{FALSE}, will not ask for confirmation before applying bonuses.}
}
\value{
Returns the WorkerIds of workers given the bonuses, or an invisible \code{NULL} if none
granted.
}
\description{
This function applies bonuses depending on whether the bonus has been received by the worker
previously for a given HITTypeId, assuming the worker achieves the threshold required by
a given qualification (or qualifications).
}
\details{
This function will grant bonuses if the worker has not received the bonus for the HITTypeId
before and if the worker has a qualification score or qualification scores that exceed the thresholds.
For example, if workers are assigned both a completion counter qualification that counts how
many of a particular HITTypeId the worker has completed, and the worker is assigned a score of
how well the worker did at completing the HITs based on gold standard or honey pot values, if the
worker exceeds a threshold on the counter (e.g., completed at least 28 assignments) AND completed
them with a score of at least 8500 points, the worker would receive the bonus.
}

