% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubMedExtractor.R
\name{MTClinicalTrials}
\alias{MTClinicalTrials}
\title{Extract Primary Outcomes from ClinicalTrials.gov}
\usage{
MTClinicalTrials(registryNumber = NULL, quiet = TRUE)
}
\arguments{
\item{registryNumber}{The registry number, in the form of NCT########}

\item{quiet}{Logical. If \code{FALSE}, additional messages should print.}
}
\value{
A string or vector of the length of \code{registryNumber} of
primary outcomes. Multiple primary outcomes for the same registryNumber
are tab delimited.
}
\description{
This method extracts data from ClinicalTrials.gov.
Essentially a specific wrapper around MTXPath.
}

