% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCountAndApprove.R
\name{MTCountAndApprove}
\alias{MTCountAndApprove}
\title{Note: non-functioning. Kept in case code snippets are still useful. Function to count assignments, update a counter qual, and approve assignments}
\usage{
MTCountAndApprove(results = NULL, counterQual = NULL, countAll = FALSE,
  feedback = "Thank you.", approve = TRUE, sandbox = TRUE,
  verbose = FALSE, outType = "full")
}
\arguments{
\item{results}{A data frame of results returned from MTurk.}

\item{counterQual}{The qualification ID string that identifies the counter qualification for this HIT.}

\item{countAll}{Logical. Determines whether the function counts all of the assignments in the
\code{results} object. If \code{FALSE} (default), the function will only count assigments with
\code{AssignmentStatus} with a value of \code{"Submitted"}.}

\item{feedback}{The text of an email to send to the workers. Default is \code{"Thank you."}.}

\item{approve}{Logical. Whether to approve assignments after counting.
This will return the \code{results} object, but with \code{AssignmentStatus}
set to \code{"ApprovedLocal"}. This prevents needing to refetch \code{results} to continue
working with the results. Default is \code{TRUE}.}

\item{sandbox}{Logical. Whether to use the sandbox (\code{TRUE}) or not; default is \code{TRUE}.}

\item{verbose}{Logical. Whether to print additional messages or not.}

\item{outType}{Either set to \code{"sub"} or \code{"full"}. If \code{"sub"},
only the newly evaluated subset will be returned. Default is \code{"full"}.}
}
\value{
If \code{approve = TRUE}, it will change the \code{AssignmentStatus} column to
\code{"ApprovedLocal"}. Otherwise, the response is silent.
}
\description{
This function fetches the count of assignments a worker has completed from MTurk,
adds a counter for newly completed assignments, and approves assignments.
}

