% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCountAssignments.R
\name{MTCountAssignments}
\alias{MTCountAssignments}
\title{Function to increment a counter of assignments}
\usage{
MTCountAssignments(results = NULL, counterQual = NULL, approve = FALSE,
  sandbox = TRUE)
}
\arguments{
\item{results}{A results object returned from MTurk.}

\item{counterQual}{The qualification ID string that identifies the counter qualification for this HIT.}

\item{approve}{Logical. Whether to approve assignments after counting. This will return the \code{results} object,
but with \code{AssignmentStatus} set to \code{ApprovedLocal}. Default is \code{FALSE}.}

\item{sandbox}{Logical. Whether to use the sandbox (\code{TRUE}) or not; default is \code{TRUE}.}
}
\description{
This function fetches the count of assignments a worker has completed from MTurk, adds a counter for newly
completed assignments, and posts the new count.
}

