% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkFunctions.R
\name{MTGetHITTypeResults}
\alias{MTGetHITTypeResults}
\title{Get results for a given HITTypeId}
\usage{
MTGetHITTypeResults(HITTypeId = NULL, reward = NULL, sandbox = TRUE)
}
\arguments{
\item{HITTypeId}{REQUIRED. The HITTypeId for the HITs to be returned.}

\item{reward}{The amount in dollars that microworkers are rewarded. This is used as a convenient means of returning
the average hourly rate.}

\item{sandbox}{Default is \code{TRUE}. Tells the functions whether to operate on the
live or sandbox Mechanical Turk sites.}
}
\value{
Returns the results of HITs that are part of the HITTypeId. Also prints the hourly rate and the
percent completed.
}
\description{
Function returns the results of a given HITTypeId.
}

