% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualFunctions.R
\name{MTGetOrInitiateQualification}
\alias{MTGetOrInitiateQualification}
\title{Helper function to get qualification values or to assign 0 if the worker does not yet have the ID}
\usage{
MTGetOrInitiateQualification(workerIds = NULL, qualId = NULL,
  sandbox = sandbox)
}
\description{
Helper function to get qualification values or to assign 0 if the worker does not yet have the ID
}

