% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{MTImport}
\alias{MTImport}
\title{Importer function for 'maker' functions}
\usage{
MTImport(f, inputLoc = NULL)
}
\arguments{
\item{f}{A file name or \code{file} connection.}

\item{inputLoc}{Relative directory for \code{f} if \code{f} is a file name. Do not add final slash to directory.}
}
\value{
A character object of the contents of the file.
}
\description{
A wrapper for \code{readLines} that pulls the file contents, with lines separated by a line return.
}

