% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPrimaryOutcomes.R
\name{MTPrimaryOutcomes}
\alias{MTPrimaryOutcomes}
\title{Extract primary outcomes}
\usage{
MTPrimaryOutcomes(registryNumber = NULL, quiet = TRUE)
}
\arguments{
\item{registryNumber}{A string or vector of registry numbers.}

\item{quiet}{Logical. If set to \code{FALSE}, additional messages
may be printed.}
}
\description{
This function loops through registryNumbers, selecting the most appropriate
registry site based on location and efficiency. This helps circumvent
depending on WHO, which is slow.
}

