% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubMedExtractor.R
\name{MTPubMed}
\alias{MTPubMed}
\title{Extract XML from PubMed}
\usage{
MTPubMed(pmids = NULL, recordContent = NULL, xpaths = NULL,
  outNames = NULL)
}
\arguments{
\item{pmids}{PubMed IDs}

\item{recordContent}{String or vector of common-name PubMed record elements, such
authors, abstract, title, and registry.}

\item{xpaths}{String or vector of actual xpaths, such as '//AccessionNumber'}

\item{outNames}{Colnames for the output object. If \code{recordContent} is
defined but \code{outNames} is undefined, the function will take use \code{
recordContent} as the column names.}
}
\description{
Extract XML from PubMed
}
\details{
Specific values are available for \code{recordContent}, and the
list will grow:
\itemize{
\item{pmid}
\item{receivedDate}
\item{acceptedDate}
\item{aheadOfPrintDate}
\item{eCollectionDate}
\item{ePublishDate}
\item{pubDate}
\item{eArticleDate}
\item{abstract}
\item{doi}
\item{registry}
}
}

