% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRejectNAs.R
\name{MTRejectNAs}
\alias{MTRejectNAs}
\title{A convenience function to reject results with NAs.}
\usage{
MTRejectNAs(results, resultsCols = NULL, extendRejects = 1,
  continue = "wait", sandbox = TRUE)
}
\arguments{
\item{results}{A results object returned by e.g. \code{MTGetHITTypeResults}. REQUIRED.}

\item{resultsCols}{The names of the columns in \code{results} to look for NAs. If \code{resultsCols} is not defined,
it defaults to taking all of the response columns in a standard \code{results} object (columns 31+).}

\item{extendRejects}{Declares how many assignments should be added to rejected HIT assignments. Default is 1.}

\item{continue}{An option of whether to proceed with rejection and extension without review. Default is 'wait', and
input values are 'Y', which will proceed without review, and 'N' which will terminate the function.}

\item{sandbox}{Default is \code{TRUE}. Defines whether to work on live or sandbox Mechanical Turk site.}
}
\value{
Returns the new rejects, unless the function is terminated which returns \code{NULL}, or if there are no new
rejects, it prints "No new assignments to reject."
}
\description{
This function rejects work that has NAs as results. This can be useful for situations where a microworker's browser
manages to circumvent required HTML tags, or when microworkers are explicitly told which questions to complete
and yet the microworker submits work with non-completed fields.
}
\details{
This function looks in the results object for non-rejected assignments, looks for NA responses, and rejects
assignments as appropriate. Prior to rejecting the HITs, the function will ask for confirmation.
}

