% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTScoreAnswers.R
\name{MTScoreAnswers}
\alias{MTScoreAnswers}
\title{Function to score known answers}
\usage{
MTScoreAnswers(results = NULL, answers = NULL, qPoints = 1,
  questionNames = NULL, scoreNAsAs = "wrong", NAValue = NULL)
}
\arguments{
\item{results}{Needs column named 'RequesterAnnotation'}

\item{answers}{Needs column named 'annotation'}

\item{qPoints}{The points assigned to each question. Either a single value or a
vector of length \code{questionNames}, where each element matches the points
assigned to each element of \code{questionNames}.}

\item{questionNames}{Columns names of questions to be compared between results
and answers. If the columns names differ between the results and answers,
\code{questionNames} can take a \code{data.frame} with colnames of 'results' and
'answers'. The rows in this data frame will be used to map the columns of results
to the columns in answers.}

\item{scoreNAsAs}{How to score NAs; possible values:
\itemize{
\item "wrong" - NAs are interpreted as wrong answers
\item "right" - NAs are interpreted as right answers
\item "value" - NAs are overwritten with the value of \code{NAValue}
}}

\item{NAValue}{The value to replace NAs with.}
}
\value{
Returns the full inputted \code{results} object appended with scores.
}
\description{
Function to score known answers
}
\examples{

#questionNames <- intersect(colnames(answers),colnames(results))
#qPoints <- c(5,5,1,5,1,5,5,5)
}

