% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTSetCred.R
\name{MTSetCred}
\alias{MTSetCred}
\title{Function to import credentials.}
\usage{
MTSetCred(filename = paste0("src/credentials.txt"))
}
\arguments{
\item{filename}{The location of the credentials file. The default is \code{"src/credentials.txt"}.}
}
\value{
NA
}
\description{
Function imports MTurk credentials from a specially formatted credentials file and executes the \code{credentials}
function of MTurkR. NOTE: credentials give access to the associated Amazon account.
Be sure if you share code that you DO NOT share the credentials file unintentionally.
}

