% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTSetup.R
\name{MTSetup}
\alias{MTSetup}
\title{Function to initiate MTurk in R}
\usage{
MTSetup()
}
\value{
Prints the live account balance associated with the credentials file.
}
\description{
Sets \code{MTurkR.sandbox} to \code{TRUE}; imports credentials with \code{MTSetCred}; and returns the live account balance.
}

