% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{MTXPath}
\alias{MTXPath}
\title{xpath support function to handle multiple returned values}
\usage{
MTXPath(doc, tag, sep = ", ", fnxn = "xmlValue")
}
\arguments{
\item{doc}{Parsed XML.}

\item{tag}{The XML tag to be searched for.}

\item{sep}{A string defining how multiple values should be separated.}

\item{fnxn}{The function to be applied in \code{xpathSapply}}
}
\value{
A string
}
\description{
Wrapper for \code{xpathSApply} to concatenate multiple xpath returned values. Also handles no returned values.
}

