% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualFunctions.R
\name{MTScoreAssignments}
\alias{MTScoreAssignments}
\title{Function to increment a counter of assignments}
\usage{
MTScoreAssignments(results = NULL, answers = NULL, scoreQual = NULL,
  howToScore = "relativeTotal", pointsPerHIT = 100, counterQual = NULL,
  pointsPerQ = 1, questionNames = NULL, scoreNAsAs = "wrong",
  approve = FALSE, sandbox = TRUE)
}
\arguments{
\item{results}{A results object returned from MTurk.}

\item{answers}{A \code{data.frame} or similar object with answers to questions in the \code{results} object.
\code{colnames} of \code{answers} must match the \code{colnames} of responses in \code{results}. For \code{results}
to be scored, it must have an annotation that matches the annotation in \code{answers}.}

\item{scoreQual}{The qualification ID string that identifies the score qualification for this HIT.}

\item{howToScore}{String with a value of \code{"runningTotal"} or \code{"relativeTotal"} (default).
If \code{"relativeTotal"}, \code{counterQual} and \code{pointsPerHIT} need to be defined.}

\item{pointsPerQ}{A number or vector of numbers of length of \code{answers}. Default is 1.}

\item{approve}{Logical. Whether to approve assignments after counting. This will return the \code{results} object,
but with \code{AssignmentStatus} set to \code{ApprovedLocal}. This prevents needing to refetch \code{results} to continue
working with the results. Default is \code{FALSE}.}

\item{sandbox}{Logical. Whether to use the sandbox (\code{TRUE}) or not; default is \code{TRUE}.}
}
\description{
This function fetches the count of assignments a worker has completed from MTurk, adds a counter for newly
completed assignments, and posts the new count.
}

