% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkFunctions.R
\name{MTsetup}
\alias{MTsetup}
\title{Function to initiate MTurk in R}
\usage{
MTsetup()
}
\value{
Prints the live account balance associated with the credentials file.
}
\description{
Sets \code{MTurkR.sandbox} to \code{TRUE}; imports credentials with \code{MTSetCred}; and returns the live account balance.
}

