% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTWhoPrimaryOutcomes.R
\name{MTWhoPrimaryOutcomes}
\alias{MTWhoPrimaryOutcomes}
\title{Extract Primary Outcomes from WHO trial database}
\usage{
MTWhoPrimaryOutcomes(registryNumber = NULL, quiet = TRUE)
}
\arguments{
\item{registryNumber}{A registry number. Various registry string types work, e.g.:
\itemize{
\item{NCT01110447}
\item{ISRCTN30870177}
}}

\item{quiet}{Logical. If \code{FALSE}, additional messages should print.}
}
\value{
Returns a named vector the length of and with names derived from
\code{registryNumber}. Multiple primary outcomes are tab-delimited.
}
\description{
This method extracts the HTML elements associated with primary outcomes in the
 WHO trial database, which also pulls in records from additional registries.
 Primary outcomes are denoted by \code{DataList12} in the HTML table. Multiple
 entries are tab delimited. \cr
 Although the WHO search allows searching multiple registries at once,
 it is slow, and often times out. Use of \code{MTClinicalTrials} is
 much faster when needed.
}
\details{
More details to come.
}

